requirements.txt
anyio==4.1.0
cffi==1.16.0
click==8.1.7
clickhouse-driver==0.2.6
duckdb==0.9.2
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
Jinja2==3.1.2
Markdown==3.5.1
MarkupSafe==2.1.3
prettytable==3.9.0
prql-python==0.10.1
psycopg==3.1.14
psycopg-binary==3.1.14
pycparser==2.21
pygit2==1.13.3
python-multipart==0.0.6
pytz==2023.3.post1
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.33.0
typing_extensions==4.9.0
tzlocal==5.2
uvicorn==0.24.0.post1
wcwidth==0.2.12
pyyaml==6.0.2
